requirements.txt
fastapi
pydantic
boto3
uvicorn
python-dotenv
